+++ Sources/KeyboardKit/Actions/KeyboardAction+Actions.swift
@@ -10,22 +10,24 @@
 import Foundation
 
 /**
- This extension defines various actions that by default will
- apply to various keyboard actions.
+ This extension defines standard gesture actions for various
+ keyboard actions.
  
  You can trigger these actions directly, but a more flexible
- way is to use an action handler. You can also choose to not
- trigger these actions at all and handle keyboard actions in
- completely custom ways.
+ approach is to use a ``KeyboardActionHandler``. The library
+ also uses these actions by default.
  */
 public extension KeyboardAction {
     
+    /**
+     This typealias represents a gesture action, which
+     */
     typealias GestureAction = (KeyboardInputViewController?) -> Void
     
     
     /**
      The action that by default should be triggered when the
-     provided `gesture` is performed on the keyboard action.
+     action is interacted with using a certain `gesture`.
      */
     func standardAction(for gesture: KeyboardGesture) -> GestureAction? {
         switch gesture {
@@ -40,13 +42,13 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is double tapped.
+     action is double tapped.
      */
     var standardDoubleTapAction: GestureAction? { nil }
     
     /**
      The action that by default should be triggered when the
-     keyboard action is long pressed.
+     action is long pressed.
      */
     var standardLongPressAction: GestureAction? {
         switch self {
@@ -57,7 +59,7 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is pressed.
+     action is pressed.
      */
     var standardPressAction: GestureAction? {
         switch self {
@@ -68,13 +70,13 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is released.
+     action is released.
      */
     var standardReleaseAction: GestureAction? { nil }
     
     /**
      The action that by default should be triggered when the
-     keyboard action is pressed, repeated until released.
+     action is pressed, and repeated until it is released.
      */
     var standardRepeatAction: GestureAction? {
         switch self {
@@ -85,7 +87,7 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is tapped.
+     action is tapped.
      */
     var standardTapAction: GestureAction? {
         if let action = standardTextDocumentProxyAction { return action }
@@ -104,33 +106,17 @@ public extension KeyboardAction {
     
     /**
      The standard text document proxy action, if any.
-     
-     Text document proxy actions affect the proxy itself and
-     not its content.
      */
     var standardTextDocumentProxyAction: GestureAction? {
-        if let action = standardTextDocumentProxyInputAction { return action }
-        switch self {
-        case .moveCursorBackward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: -1) }
-        case .moveCursorForward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: 1) }
-        default: return nil
-        }
-    }
-    
-    /**
-     The standard text document proxy input action, if any.
-     
-     Text document proxy input actions affect the text proxy
-     content, e.g. inserting or deleting text.
-     */
-    var standardTextDocumentProxyInputAction: GestureAction? {
-        if self.isPrimaryAction { return { $0?.textDocumentProxy.insertText(.newline) }}
         switch self {
         case .backspace: return { $0?.textDocumentProxy.deleteBackward(range: $0?.keyboardBehavior.backspaceRange ?? .char) }
         case .character(let char): return { $0?.textDocumentProxy.insertText(char) }
         case .characterMargin(let char): return { $0?.textDocumentProxy.insertText(char) }
         case .emoji(let emoji): return { $0?.textDocumentProxy.insertText(emoji.char) }
+        case .moveCursorBackward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: -1) }
+        case .moveCursorForward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: 1) }
         case .newLine: return { $0?.textDocumentProxy.insertText(.newline) }
+        case .primary: return { $0?.textDocumentProxy.insertText(.newline) }
         case .return: return { $0?.textDocumentProxy.insertText(.newline) }
         case .space: return { $0?.textDocumentProxy.insertText(.space) }
         case .tab: return { $0?.textDocumentProxy.insertText(.tab) }