+++ RELEASE_NOTES.md
@@ -28,7 +28,9 @@ This version adds new locales and makes the library build on more platforms.
 * All previously deprecated code has been removed.
 * `KeyboardAction` `isKeyboardType` has been removed.
 * `KeyboardAction` `isSpace` has been removed.
+* `KeyboardAction` `standardTextDocumentProxyInputAction` has been merged with `standardTextDocumentProxyAction`.
 * `KeyboardInputViewController` `needsInputModeSwitchKeyOverride` has been removed.
+* `StandardKeyboardActionHandler` `GestureAction` has been removed.